requirements.txt
PySide6==6.6.1
requests==2.31.0
python-dotenv==1.0.0
pyaudio==0.2.13
numpy==1.26.2 